Dat/dev

+++ controllers/carts.py
@@ -1,22 +1,46 @@
+from notice import Notice
 from form_action import FormAction
+from frontend_view import FrontendView
+class Carts:
+    def index(self):
+        view_process = FrontendView()
+        form_process = FormAction()
+        button = view_process.show_buttons(
+            list_button=[
+                [URL('web2py_shop','orders','index'),"Purchased"],
+                [URL('web2py_shop','shop','view'),"Home"]
+            ]
+        )
+        auth_creater = auth.has_membership('create_user')
+        if auth_creater:
+            rows = db((db.auth_user.id==db.carts.user_id) & (db.carts.shop_id == db.shop.id)).select()
+        else:
+            rows = db((db.auth_user.id==db.carts.user_id) & (db.carts.shop_id == db.shop.id) & (db.carts.user_id == auth.user.id)).select()
+
+        list_id = [x.cart_id for x in db(db.orders).select(db.orders.cart_id)]
+        rows = form_process.check_purchased_row(table=rows, list_id=list_id)
+        th_list =["Shop Item","Num"]
+        column_list=["carts.id","shop.shop_img","shop.shop_item","carts.num", "show_user","purchase_btn","delete_btn"]
+        table = view_process.show_table(
+            th_list=th_list,
+            column_list=column_list,
+            table=rows,
+            is_cart=True,
+            permission= auth_creater
+        )
+
+        form = view_process.display_body(
+            head="",
+            button=button,
+            content=table
+        )
+        return form
 
 @auth.requires_login()
 def index():
-    shop_dict = []
-    user_dict = []
-    auth_creater = auth.has_membership('create_user')
-    try:
-        if auth_creater:
-            rows = db(db.carts).select()
-        else:
-            rows = db(db.carts.user_id == auth.user.id).select()
-        shops = db(db.shop).select()
-        shop_dict = {row.id: (row.shop_item, row.shop_img, row.num) for row in shops}
-        user_rows = db(db.auth_user).select()
-        user_dict = {x.id: f"{x.last_name} {x.first_name}" for x in user_rows}
-    except Exception as e:
-        rows = []
-    return dict(rows=rows, shop_dict=shop_dict, auth_creater=auth_creater, user_dict=user_dict)
+    cart_show = Carts()
+    form = cart_show.index()
+    return dict(form=form)
 
 @auth.requires_login()
 def post():
@@ -26,22 +50,34 @@ def post():
             quantity = int(data.quantity)
             shop_id = int(data.shop_id)
             user_id = int(auth.user.id)
-            check_row = db((db.carts.shop_id == shop_id) & (db.carts.user_id == user_id)).select().first()
+            check_row = db((db.carts.shop_id == shop_id) & (db.carts.user_id == user_id)).select()
+            
+            form_process = FormAction()
+            list_id = [x.cart_id for x in db(db.orders).select(db.orders.cart_id)]
+            check_row = form_process.check_purchased_row(table=check_row,list_id=list_id)
+            notice_process = Notice()
+            notice_process.mail_item_notice(
+                mail_to_list=['nguyen321lht@gmail.com'],
+                action="add",
+                item=db.shop(shop_id).shop_item,
+                num=quantity
+            )
+
             if check_row:
-                id = int(check_row.id)
-                redirect((URL('demo','provider','form', vars=dict(id=id,num_change=quantity))))
+                id = int(check_row[0].id)
+                redirect((URL('web2py_shop','provider','form', vars=dict(id=id,num_change=quantity))))
             current_num = db.shop(shop_id).num  
             new_num = current_num - quantity
             db.carts.insert(shop_id=shop_id, user_id=user_id, num=quantity)
             db(db.shop.id == shop_id).update(num=new_num)
-            redirect(URL('demo','carts','index'))
+            redirect(URL('web2py_shop','carts','index'))
         except Exception as e:
             print(e)
             if check_row:
-                redirect(URL('demo','carts','update', vars=dict(id=id,num_change=quantity)))
-            redirect(URL('demo','carts','index'))
+                redirect(URL('web2py_shop','carts','update', vars=dict(id=id,num_change=quantity)))
+            redirect(URL('web2py_shop','carts','index'))
     else:
-        redirect(URL('demo','carts','index'))
+        redirect(URL('web2py_shop','carts','index'))
 
 @auth.requires_login()
 def delete():
@@ -49,23 +85,33 @@ def delete():
         try:
             data = request.vars
             id = int(data.id)
-            if auth.has_membership('create_user') or db.carts(id).user_id == auth.user.id:
+            list_id = [x.cart_id for x in db(db.orders).select(db.orders.cart_id)]
+            if (auth.has_membership('create_user') or db.carts(id).user_id == auth.user.id) and id not in list_id:
                 shop_id = db.carts(id).shop_id
                 current_num_carts = db.carts(id).num
                 current_num_shop = db.shop(shop_id).num
                 new_num_shop = current_num_shop + current_num_carts
                 db(db.shop.id == shop_id).update(num=new_num_shop)
                 db(db.carts.id == id).delete()
-                redirect(URL('demo','carts','index'))
+
+                notice_process = Notice()
+                notice_process.mail_item_notice(
+                    mail_to_list=['nguyen321lht@gmail.com'],
+                    action="return",
+                    item=db.shop(shop_id).shop_item,
+                    num=current_num_carts
+                )
+                redirect(URL('web2py_shop','carts','index'))
         except Exception as e:
             print(e)
-            redirect(URL('demo','carts','index'))
+            redirect(URL('web2py_shop','carts','index'))
     else:
-        redirect(URL('demo','carts','index'))
+        redirect(URL('web2py_shop','carts','index'))
 
 @auth.requires_login()
 def update():
     try:
+        list_id = [x.cart_id for x in db(db.orders).select(db.orders.cart_id)]
         data = request.vars
         num_change = int(data.num_change)
         id = int(data.id)
@@ -74,10 +120,12 @@ def update():
         current_num_shop = db.shop(shop_id).num
         new_num_cart = current_num_carts+num_change
         new_num_shop = current_num_shop-num_change
-        if new_num_cart>0 or new_num_shop >=0:
+        if (new_num_cart>0 or new_num_shop >=0) and (
+                auth.has_membership('create_user') or db.carts(id).user_id == auth.user.id
+            ) and id not in list_id:
             db(db.shop.id == shop_id).update(num = new_num_shop)
             db(db.carts.id == id).update(num = new_num_cart)
     except Exception as e:
         print(e)
-        redirect(URL('demo','carts','index'))
-    return redirect(URL('demo','carts','index'))
+        redirect(URL('web2py_shop','carts','index'))
+    return redirect(URL('web2py_shop','carts','index'))
